Create BaseDataTable class Reduce code duplication.
