+++ app/DataTables/Alerting/LogsDataTable.php
@@ -25,10 +25,10 @@
 
 namespace App\DataTables\Alerting;
 
+use App\DataTables\BaseDataTable;
 use App\Models\Alerting\Log;
-use Yajra\Datatables\Services\DataTable;
 
-class LogsDataTable extends DataTable
+class LogsDataTable extends BaseDataTable
 {
     /**
      * Display ajax response.
@@ -100,23 +100,11 @@ class LogsDataTable extends DataTable
     }
 
     /**
-     * Optional method if you want to use html builder.
-     *
-     * @return \Yajra\Datatables\Html\Builder
-     */
-    public function html()
-    {
-        return $this->builder()
-                    ->columns($this->getColumns())
-                    ->parameters($this->getBuilderParameters());
-    }
-
-    /**
      * Get columns.
      *
      * @return array
      */
-    private function getColumns()
+    public function getColumns()
     {
         return [
             'device.hostname' => [
@@ -139,27 +127,6 @@ class LogsDataTable extends DataTable
      */
     protected function filename()
     {
-        return 'logs';
+        return 'alert_logs';
     }
-
-    /**
-     * Get Builder Params
-     *
-     * @return array
-     */
-    protected function getBuilderParameters()
-    {
-        return [
-            'dom' => "<'row'<'col-sm-3'l><'col-sm-6 text-center'B><'col-sm-3'f>>".
-                "<'row'<'col-sm-12'tr>>".
-                "<'row'<'col-sm-5'i><'col-sm-7'p>>",
-            'order' => [3, 'desc'],
-            'lengthMenu' => [[25, 50, 100, -1], [25, 50, 100, "All"]],
-            'buttons' => [
-                'csv', 'excel', 'pdf', 'print', 'reset', 'reload',
-            ],
-            'autoWidth' => false,
-        ];
-    }
-
 }
