+++ app/DataTables/BaseDataTable.php
@@ -1,0 +1,77 @@
+<?php
+/**
+ * BaseDataTable.php
+ *
+ * -Description-
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @package    LibreNMS
+ * @link       http://librenms.org
+ * @copyright  2016 Tony Murray
+ * @author     Tony Murray <murraytony@gmail.com>
+ */
+
+namespace app\DataTables;
+
+
+use Yajra\Datatables\Services\DataTable;
+
+abstract class BaseDataTable extends DataTable
+{
+    /**
+     * Optional method if you want to use html builder.
+     *
+     * @return \Yajra\Datatables\Html\Builder
+     */
+    public function html()
+    {
+        return $this->builder()
+            ->columns($this->getColumns())
+            ->parameters($this->getBuilderParameters());
+    }
+
+    /**
+     * Get columns.
+     *
+     * @return array
+     */
+    abstract public function getColumns();
+
+    /**
+     * Get the query object to be processed by datatables.
+     *
+     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     */
+    abstract public function query();
+    
+    /**
+     * Get Builder Params
+     *
+     * @return array
+     */
+    protected function getBuilderParameters()
+    {
+        return [
+            'dom' => "<'row'<'col-sm-3'l><'col-sm-6 text-center'B><'col-sm-3'f>>".
+                "<'row'<'col-sm-12'tr>>".
+                "<'row'<'col-sm-5'i><'col-sm-7'p>>",
+            'lengthMenu' => [[25, 50, 100, -1], [25, 50, 100, "All"]],
+            'buttons' => [
+                'csv', 'excel', 'pdf', 'print', 'reset', 'reload',
+            ],
+            'autoWidth' => false,
+        ];
+    }
+}
