+++ app/DataTables/DeviceDataTable.php
@@ -26,9 +26,8 @@
 namespace App\DataTables;
 
 use App\Models\Device;
-use Yajra\Datatables\Services\DataTable;
 
-class DeviceDataTable extends DataTable
+class DeviceDataTable extends BaseDataTable
 {
     /**
      * Display ajax response.
@@ -86,23 +85,11 @@ class DeviceDataTable extends DataTable
     }
 
     /**
-     * Optional method if you want to use html builder.
-     *
-     * @return \Yajra\Datatables\Html\Builder
-     */
-    public function html()
-    {
-        return $this->builder()
-                    ->columns($this->getColumns())
-                    ->parameters($this->getBuilderParameters());
-    }
-
-    /**
      * Get columns.
      *
      * @return array
      */
-    private function getColumns()
+    public function getColumns()
     {
         return [
             'status'    => [
@@ -148,24 +135,4 @@ class DeviceDataTable extends DataTable
     {
         return 'devices';
     }
-
-    /**
-     * Get Builder Params
-     *
-     * @return array
-     */
-    protected function getBuilderParameters()
-    {
-        return [
-            'dom' => "<'row'<'col-sm-3'l><'col-sm-6 text-center'B><'col-sm-3'f>>".
-                     "<'row'<'col-sm-12'tr>>".
-                     "<'row'<'col-sm-5'i><'col-sm-7'p>>",
-            'lengthMenu' => [[25, 50, 100, -1], [25, 50, 100, "All"]],
-            'buttons' => [
-                'csv', 'excel', 'pdf', 'print', 'reset', 'reload',
-            ],
-            'autoWidth' => false,
-        ];
-    }
-
 }
