+++ app/DataTables/General/SearchDataTable.php
@@ -25,13 +25,13 @@
 
 namespace App\DataTables\General;
 
+use App\DataTables\BaseDataTable;
 use App\Models\General\IPv4;
 use App\Models\General\IPv4Mac;
 use App\Models\General\IPv6;
 use App\Models\Port;
-use Yajra\Datatables\Services\DataTable;
 
-class SearchDataTable extends DataTable
+class SearchDataTable extends BaseDataTable
 {
 
     protected $type;
@@ -117,23 +117,11 @@ class SearchDataTable extends DataTable
     }
 
     /**
-     * Optional method if you want to use html builder.
-     *
-     * @return \Yajra\Datatables\Html\Builder
-     */
-    public function html()
-    {
-        return $this->builder()
-                    ->columns($this->getColumns())
-                    ->parameters($this->getBuilderParameters());
-    }
-
-    /**
      * Get columns.
      *
      * @return array
      */
-    private function getColumns()
+    public function getColumns()
     {
         $cols = [
             'hostname'  => [
@@ -203,24 +191,6 @@ class SearchDataTable extends DataTable
     protected function filename()
     {
         return 'search';
-    }
-
-    /**
-     * Get Builder Params
-     *
-     * @return array
-     */
-    protected function getBuilderParameters()
-    {
-        return [
-            'dom' => "<'row'<'col-sm-3'l><'col-sm-6 text-center'B><'col-sm-3'f>>".
-                "<'row'<'col-sm-12'tr>>".
-                "<'row'<'col-sm-5'i><'col-sm-7'p>>",
-            'lengthMenu' => [[25, 50, 100, -1], [25, 50, 100, "All"]],
-            'buttons' => [
-                'csv', 'excel', 'pdf', 'print', 'reset', 'reload',
-            ],
-        ];
     }
